requirements.txt
Flask==2.3.2
MarkupSafe==2.1.1
Pillow==10.3.0
python-dotenv==1.0.1
google-generativeai==0.5.2